+++ project/src/main/java/com/betacom/businesscomponent/idgenerator/IdGenerator.java
@@ -5,5 +5,5 @@ import java.io.IOException;
 import com.betacom.architecture.dao.DAOException;
 
 public interface IdGenerator {
-long getNextId() throws ClassNotFoundException, DAOException, IOException; 
+int getNextId() throws ClassNotFoundException, DAOException, IOException; 
 }
